backend/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
requests==2.32.3
dotenv-python==0.0.1
openai==1.70.0
django-cors-headers==4.7.0
django-model-utils==5.0.0
djangorestframework-simplejwt==5.5.0
djangorestframework-camel-case==1.4.2
drf-spectacular[sidecar]==0.28.0
django-filter==24.3
psycopg==3.2.4
psycopg2-binary==2.9.10
django_redis==5.4.0
celery==5.5.0
django-celery-results==2.5.1
flower==2.0.1



